Add files via upload
This branch contains the evaluated notebook with outputs included. The file has not been edited, except for the removal of simulation configurations and some code blocks unrelated to the main content. These removed blocks were auxiliary verifications not directly relevant to the content shared and do not affect the execution of the notebook. The removal was made solely to reduce the file size and ensure compatibility with the official version published on GitHub.

Notes:

The notebook is in Portuguese.
It does not have the same level of formatting and adjustments as the official version published in the article.uts for verification.